+++ packages/client/components/UserDraggingHeader.tsx
@@ -1,4 +1,4 @@
-import {keyframes} from '@emotion/core'
+import {keyframes} from '@emotion/react'
 import styled from '@emotion/styled'
 import {ArrowBack, ArrowDownward, ArrowForward, ArrowUpward} from '@mui/icons-material'
 import * as React from 'react'