+++ packages/server/postgres/migrations/2025-01-08T15:10:42.442Z_addRetroTemplateLikeWishWonder.ts
@@ -1,5 +1,5 @@
 import type {Kysely} from 'kysely'
-import getTemplateIllustrationUrl from 'parabol-server/graphql/mutations/helpers/getTemplateIllustrationUrl'
+import getTemplateIllustrationUrl from '../../graphql/mutations/helpers/getTemplateIllustrationUrl'
 
 export async function up(db: Kysely<any>): Promise<void> {
   await db