+++ scripts/runSocketServer.js
@@ -7,6 +7,7 @@
 try {
   require('../dev/web.js')
 } catch (e) {
+  console.error(e)
   // webpack has not created the file yet
   // pm2 will restart this process when the file changes
 }